action: add check-files-from input to specify files to check
This commit adds a new input parameter `check-files-from` to the GitHub
Action configuration. This parameter allows users to specify a file
containing a list of files to be checked for correct formatting; it is
especially useful to limit the check to the files being touched in the
PR.

When this parameter is set, it overrides the existing `check-path`
parameter.